+++ packages/blocks/snowflake/src/lib/common/configure-connection.ts
